+++ Cargo.toml
@@ -19,5 +19,9 @@ bitflags = "1.0"
 unsafe_unwrap = "0.1"
 
 [dev-dependencies]
+criterion = "0.2"
 rand = "0.6"
 
+[[bench]]
+name = "criterion"
+harness = false
